move Metric model to indexer
nothing in storage driver uses Metric.

Closes: #364
